+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - made `InitError` visible
 - added `ILI9488` model support
 - added `KIND` constant to `Interface` trait to detect invalid model, color format, and interface combinations
+- added `InvalidConfiguration` variant to `InitError` enum
+
+### Changed
+
+- changed the returned error type of `Model::init` to a new `ModelInitError` type to allow implementations to report configuration errors
 
 ## Removed